👌 IMPROVE: Created getAPIListData utility
After yesterdays demo, I realized the site was not dynamic when it should be. This PR corrects that issue.

A new utility file was created to bring the dynamic functionality to the rest of the site.
