+++ server/routes/base-apis.js
@@ -1,29 +1,34 @@
 const express = require("express");
+const router = express.Router();
 const path = require("path");
 const jsonServer = require("json-server");
 const jsonGraphqlExpress = require("json-graphql-server");
+
+const { getAPIListData } = require("../utils/getAPIListData");
 
 const { apiLimits } = require("../utils/rateLimiterDefaults");
 const { getFromFile } = require("../utils/utils");
 
 const { verifyData } = require("../utils/verifyData");
 
-const ApiList = require("../apiList");
+const init = async () => {
+  const APIListData = await getAPIListData();
 
-const router = express.Router();
+  APIListData.forEach(({ link }) => {
+    const dataPath = path.join(__dirname, `../api/${link}.json`);
+    const data = getFromFile(dataPath);
 
-// API EndPoint Route
-ApiList.forEach(({ link }) => {
-  const dataPath = path.join(__dirname, `../api/${link}.json`);
-  const data = getFromFile(dataPath);
-  try {
-    router.use(`/${link}/graphql`, apiLimits, jsonGraphqlExpress.default(data));
-  } catch (err) {
-    console.log(`Unable to set up /${link}/graphql`);
-    console.error(err);
-  }
+    try {
+      router.use(`/${link}/graphql`, apiLimits, jsonGraphqlExpress.default(data));
+    } catch (err) {
+      console.log(`Unable to set up /${link}/graphql`);
+      console.error(err);
+    }
 
-  router.use(`/${link}`, verifyData, apiLimits, jsonServer.router(dataPath));
-});
+    router.use(`/${link}`, verifyData, apiLimits, jsonServer.router(dataPath));
+  });
+};
+
+init();
 
 module.exports = router;
