+++ server/routes/frontend.js
@@ -1,28 +1,26 @@
 const express = require("express");
-const fs = require("fs");
-const path = require("path");
-
 const router = express.Router();
 
-let APIList = [];
-const directoryPath = path.join(__dirname, "../api");
+const { getAPIListData } = require("../utils/getAPIListData");
+
+let APIListData = [];
 
 router.get("/", async (req, res) => {
-  if (!APIList.length) {
-    await getApiList();
+  if (!APIListData.length) {
+    APIListData = await getAPIListData();
   }
 
   res.json({
     status: 200,
     data: {
-      APIList,
+      APIListData,
     },
   });
 });
 
 router.get("/:name", async (req, res) => {
-  if (!APIList.length) {
-    await getApiList();
+  if (!APIListData.length) {
+    APIListData = await getAPIListData();
   }
 
   res.json({
@@ -30,26 +28,5 @@ router.get("/:name", async (req, res) => {
     id: req.params.name,
   });
 });
-
-const getApiList = async () => {
-  let files = await fs.readdirSync(directoryPath);
-  files = files.filter((f) => !f.includes(".backup"));
-  APIList = files.map((file) => {
-    if (file.includes(".backup")) {
-      return false;
-    }
-    const data = JSON.parse(fs.readFileSync(path.join(__dirname, `../api/${file}`)));
-    const endpoints = Object.keys(data);
-    const metaData = data.metaData[0];
-    const name = file.split(".")[0].toLowerCase();
-
-    return {
-      name,
-      link: name,
-      metaData,
-      endpoints: endpoints.filter((e) => e != "metaData"),
-    };
-  });
-};
 
 module.exports = router;
